Add bird class with all bird functions
