+++ lib/index.js
@@ -1,22 +1,18 @@
-var birdCanvas = document.getElementById("birdWindow")
+
 var playerInput = document.getElementById("player-input")
 var startButton = document.getElementById("start-button")
 var counter = 0;
-var birdPoints = 0;
+// var birdPoints = 0;
 var indexCounter = 0;
-var birdContext = birdCanvas.getContext("2d");
+// var birdContext = birdCanvas.getContext("2d");
 var previousCharacter = null;
 
-const Scorecard = require('./scorecard')
+const Scorecard = require('./scorecard');
+const Bird = require('./bird')
 var scorecard = new Scorecard();
+var bird = new Bird();
 
 
-function bird() {
-	this.x = 10;
-	this.y= 150;
-	this.width = 10;
-	this.height = 10;
-}
 
 function levelText(string) {
 	this.text = string
@@ -35,11 +31,10 @@ function keyStroke(char, challengeText) {
 	this.score = function() {
 		if (this.char === challengeText[indexCounter]) {
 			scorecard.birdPoints++;
-			console.log(scorecard.birdPoints)
 			counter++;
 			indexCounter++;
-			myBird.up();
-			// myBird.move().draw();
+			bird.up();
+			// bird.move().draw();
 			this.correct = true;
 			return this;
 		} else if (this.char === "Backspace"){
@@ -49,7 +44,7 @@ function keyStroke(char, challengeText) {
 		} else {
 				scorecard.birdPoints--;
 				counter++;
-				myBird.down();
+				bird.down();
 				this.correct = false;
 				return this;
 		};
@@ -78,39 +73,15 @@ function keyStroke(char, challengeText) {
 	}
 }
 
-var myBird = new bird();
-
-bird.prototype.fall = function(){
-	var y = this.y
-	var x = this.x
-	requestAnimationFrame(function gameLoop () {
-		birdContext.fillStyle ="black";
-		birdContext.clearRect(0, 0, 300, 300);
-		birdContext.fillRect(50, y+=0.25, 10, 10);
-		requestAnimationFrame(gameLoop);
-	})
-	return this;
-}
-
-bird.prototype.up = function(){
-	this.y-=15;
-	return this;
-}
-
-bird.prototype.down = function(){
-	this.y+=15;
-	return this;
-}
-
 startButton.addEventListener('click', function(){
 	var levelOne = new levelText("hello world");
 	levelOne.render();
-	myBird.fall();
+	bird.fall();
 	this.remove();
 })
 
 playerInput.addEventListener('keyup', function keyStore(event){
-	myBird.fall(this.y);
+	bird.fall(this.y);
 	var challengeText = $('#challenge-text').text()
 	var currentKey = event.key;
 	var stroke = new keyStroke(currentKey, challengeText)
